requirements.txt
joblib==1.2.0
requests==2.28.1
scikit-learn==1.0.2
numpy==1.23.0
shap==0.41.0
xgboost==1.7.3
plotly==5.15.0
streamlit==1.17.0
requests==2.28.1
Pillow==10.0.0
altair==4.0.0
matplotlib==3.5.2
numpy==1.23.0
